Add function to read pull request template
This commit adds a new function called `findPullRequestTemplate()` to the `src/utils/readPullRequestTemplate.ts` file. This function searches for a `pull_request_template.md` file in the `.github` directory of the current working directory or its ancestors. If found, the function reads the contents of the template file as a UTF-8 encoded string and returns it. If the template file is not found, the function returns `null`.

This new function will be useful for automating the creation of pull request descriptions by allowing us to include a standard template that can be easily customized for each pull request.